requirements.txt
Flask
Gunicorn
langchain
google-api-python-client
openai
python-dotenv
requests
pyjwt